+++ tests/play.test.ts
@@ -45,7 +45,7 @@ test.describe('Play', () => {
 			const generateButton = page.locator('button.chat__button-generate');
 			await expect(generateButton).toBeDisabled();
 
-			let prompt = MockPrompt.GENERATE_ARTICLE;
+			let prompt = MockPrompt.GENERATE_ARTICLE_0;
 			await page.locator('textarea').fill(prompt);
 			await expect(generateButton).not.toBeDisabled();
 			await expect(page.getByText(MOCK_ARTICLE_COMPLETIONS[0].category)).not.toBeVisible();
@@ -65,11 +65,11 @@ test.describe('Play', () => {
 
 			let article = await getLastArticle(`headline = "${MOCK_ARTICLE_COMPLETIONS[0].headline}"`);
 			let articleMessages = article?.expand?.['messages(article)'];
-			expect(articleMessages?.[0].content).toBe(MockPrompt.GENERATE_ARTICLE);
+			expect(articleMessages?.[0].content).toBe(MockPrompt.GENERATE_ARTICLE_0);
 			expect(articleMessages?.[1].content?.notes).toBe(MOCK_ARTICLE_COMPLETIONS[0].notes);
 			expect(article?.status).toBe(ArticleStatus.DRAFT);
 
-			prompt = MockPrompt.RETRY_ARTICLE;
+			prompt = MockPrompt.GENERATE_ARTICLE_1;
 			await page.locator('textarea').fill(prompt);
 			await expect(page.getByText(MOCK_ARTICLE_COMPLETIONS[1].category)).not.toBeVisible();
 			await expect(page.getByText(MOCK_ARTICLE_COMPLETIONS[1].headline)).not.toBeVisible();
@@ -88,7 +88,7 @@ test.describe('Play', () => {
 
 			article = await getLastArticle(`headline = "${MOCK_ARTICLE_COMPLETIONS[1].headline}"`);
 			articleMessages = article?.expand?.['messages(article)'];
-			expect(articleMessages?.[2].content).toBe(MockPrompt.RETRY_ARTICLE);
+			expect(articleMessages?.[2].content).toBe(MockPrompt.GENERATE_ARTICLE_1);
 			expect(articleMessages?.[3].content?.notes).toBe(MOCK_ARTICLE_COMPLETIONS[1].notes);
 			expect(article?.status).toBe(ArticleStatus.DRAFT);
 		});
@@ -97,7 +97,7 @@ test.describe('Play', () => {
 			// First prompt
 			const generateButton = page.locator('button.chat__button-generate');
 			const publishButton = page.locator('button.form-button', { hasText: 'Publish' });
-			await page.locator('textarea').fill(MockPrompt.GENERATE_ARTICLE);
+			await page.locator('textarea').fill(MockPrompt.GENERATE_ARTICLE_0);
 			await expect(page.getByText(MOCK_ARTICLE_COMPLETIONS[0].headline)).not.toBeVisible();
 			await expect(publishButton).not.toBeVisible();
 
@@ -107,7 +107,7 @@ test.describe('Play', () => {
 
 			// Second prompt
 			await expect(page.getByText(MOCK_ARTICLE_COMPLETIONS[1].headline)).not.toBeVisible();
-			await page.locator('textarea').fill(MockPrompt.RETRY_ARTICLE);
+			await page.locator('textarea').fill(MockPrompt.GENERATE_ARTICLE_1);
 
 			await generateButton.click();
 			await expect(page.getByText(MOCK_ARTICLE_COMPLETIONS[1].headline)).toBeVisible();
@@ -120,7 +120,7 @@ test.describe('Play', () => {
 			await expect(publishButton).not.toBeVisible();
 
 			let article = await getLastArticle(`headline = "${MOCK_ARTICLE_COMPLETIONS[1].headline}"`);
-			expect(article?.expand?.['messages(article)']?.[2].content).toBe(MockPrompt.RETRY_ARTICLE);
+			expect(article?.expand?.user.email).toBe(MOCK_USERS.alice.email);
 			expect(article?.status).toBe(ArticleStatus.PUBLISHED);
 
 			article = await getLastArticle(`headline = "${MOCK_ARTICLE_COMPLETIONS[0].headline}"`);
@@ -131,33 +131,33 @@ test.describe('Play', () => {
 			// First prompt
 			const generateButton = page.locator('button.chat__button-generate');
 			const publishButton = page.locator('button.form-button', { hasText: 'Publish' });
-			await page.locator('textarea').fill(MockPrompt.GENERATE_ARTICLE);
-			await expect(page.getByText(MOCK_ARTICLE_COMPLETIONS[0].headline)).not.toBeVisible();
+			await page.locator('textarea').fill(MockPrompt.GENERATE_ARTICLE_2);
+			await expect(page.getByText(MOCK_ARTICLE_COMPLETIONS[2].headline)).not.toBeVisible();
 			await expect(publishButton).not.toBeVisible();
 
 			await generateButton.click();
-			await expect(page.getByText(MOCK_ARTICLE_COMPLETIONS[0].headline)).toBeVisible();
+			await expect(page.getByText(MOCK_ARTICLE_COMPLETIONS[3].headline)).not.toBeVisible();
+			await expect(page.getByText(MOCK_ARTICLE_COMPLETIONS[2].headline)).toBeVisible();
 			await expect(publishButton).toBeVisible();
 
 			// Second prompt
-			await expect(page.getByText(MOCK_ARTICLE_COMPLETIONS[1].headline)).not.toBeVisible();
-			await page.locator('textarea').fill(MockPrompt.RETRY_ARTICLE);
-
+			await page.locator('textarea').fill(MockPrompt.GENERATE_ARTICLE_3);
 			await generateButton.click();
-			await expect(page.getByText(MOCK_ARTICLE_COMPLETIONS[1].headline)).toBeVisible();
+			await expect(page.getByText(MOCK_ARTICLE_COMPLETIONS[2].headline)).toBeVisible();
+			await expect(page.getByText(MOCK_ARTICLE_COMPLETIONS[3].headline)).toBeVisible();
 
 			// Publish article generated by second prompt
 			await publishButton.first().click();
 			await expect(page.locator('h1.section__h1', { hasText: MOCK_USERS.alice.nickname })).toBeVisible(); // prettier-ignore
-			await expect(page.getByText(MOCK_ARTICLE_COMPLETIONS[0].headline)).toBeVisible();
+			await expect(page.getByText(MOCK_ARTICLE_COMPLETIONS[2].headline)).toBeVisible();
 			await expect(generateButton).not.toBeVisible();
 			await expect(publishButton).not.toBeVisible();
 
-			let article = await getLastArticle(`headline = "${MOCK_ARTICLE_COMPLETIONS[0].headline}"`);
-			expect(article?.expand?.['messages(article)']?.[2].content).toBe(MockPrompt.RETRY_ARTICLE);
+			let article = await getLastArticle(`headline = "${MOCK_ARTICLE_COMPLETIONS[2].headline}"`);
+			expect(article?.expand?.user.email).toBe(MOCK_USERS.alice.email);
 			expect(article?.status).toBe(ArticleStatus.PUBLISHED);
-
-			article = await getLastArticle(`headline = "${MOCK_ARTICLE_COMPLETIONS[1].headline}"`);
+			
+			article = await getLastArticle(`headline = "${MOCK_ARTICLE_COMPLETIONS[3].headline}"`);
 			expect(article).toBe(null);
 		});
 
@@ -242,7 +242,7 @@ test.describe('Play', () => {
 			await expect(startFromScratchButton).not.toBeVisible();
 			await expect(publishButton).not.toBeVisible();
 
-			let prompt = MockPrompt.GENERATE_ARTICLE;
+			let prompt = MockPrompt.GENERATE_ARTICLE_0;
 			await page.locator('textarea').fill(prompt);
 			await generateButton.click();
 			await expect(page.getByText(MOCK_ARTICLE_COMPLETIONS[0].headline)).toBeVisible();
@@ -250,7 +250,7 @@ test.describe('Play', () => {
 			await expect(startFromScratchButton).toBeVisible();
 			await expect(publishButton).toBeVisible();
 
-			prompt = MockPrompt.RETRY_ARTICLE;
+			prompt = MockPrompt.GENERATE_ARTICLE_1;
 			await page.locator('textarea').fill(prompt);
 			await generateButton.click();
 			await expect(page.getByText(MOCK_ARTICLE_COMPLETIONS[0].headline)).toBeVisible();
